tried to connect PC using private IP from same LAN

+++ app/src/main/java/com/example/weatherwear/data/model/Weather.kt
@@ -1,5 +1,5 @@
 package com.example.weatherwear.data.model
 
 data class Weather (
-    val temperature : Int
+    val temperature : Double
 )
